make constructor of LatLngBounds public, rename package org.apidesign.html.leaflet.api.event and org.apidesign.html.leaflet.api.listener, implement add event listeners in Map and TileLayer, tests in Main
