+++ api/src/main/java/org/apidesign/html/leaflet/api/listener/MouseListener.java
@@ -24,16 +24,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.apidesign.html.leaflet.api.map.listener;
+package org.apidesign.html.leaflet.api.listener;
 
+import org.apidesign.html.leaflet.api.event.MouseEvent;
 import java.util.EventListener;
-import org.apidesign.html.leaflet.api.map.event.TileEvent;
 
 /**
  *
  * @author Andreas Grimmer
  */
-public interface TileListener extends EventListener {
+public interface MouseListener extends EventListener {
     
-    public void onEvent(TileEvent ev);
+    public void onEvent(MouseEvent ev);
 }
